requirements.txt
scrapy==2.12.0
pymongo 
mysql-connector-python
#lxml==5.3.0
#Scrapy